% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{arma}
\alias{arma}
\alias{AR}
\alias{MA}
\title{Control the autorregresive and moving average components}
\usage{
AR(epidemic_trend = 1, delay_trend = 1)

MA(epidemic_errors = 1)
}
\arguments{
\item{epidemic_trend}{Integer. Degree of the epidemic trend. Refers to the p parameter in an ARMA(p,q)
for the epidemic process.}

\item{delay_trend}{Integer. Degree of the delay trend. Refers to the q parameter
in an ARMA(p,q) for the delay process.}

\item{epidemic_errors}{Integer. Degree of the errors in the epidemic trend. Refers to the q parameter
in an ARMA(p,q) for the epidemic process.}
}
\description{
Controls either the autoregresive \code{\link[=AR]{AR()}} or the moving-average \code{\link[=MA]{MA()}} components
of the epidemic and delay process
}
\examples{
AR(2,2)
MA(1)

}
