% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_to_list}
\alias{array_to_list}
\title{Transforms an array into a list of lists}
\usage{
array_to_list(my_array, last_dim_as = "vector")
}
\arguments{
\item{my_array}{The array to transform}

\item{last_dim_as}{Either 'vector', 'matrix'  or 'scalar'  depending on what
we want for the last dimension}
}
\value{
A list of lists with the same structure as the array
}
\description{
Function that takes an array and transforms it into lists of lists
this is mainly for interacting with [Rcpp::cppFunction()].
}
\keyword{internal}
