% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{calc_quantile_scores}
\alias{calc_quantile_scores}
\title{calc_quantile_scores}
\usage{
calc_quantile_scores(backtest_summary, metrics)
}
\arguments{
\item{backtest_summary}{results of backtest call}

\item{metrics}{the quantile metrics to calculate - see \code{\link{get_quantile_metrics}}}
}
\value{
The weighted interval score for each of the runs in \code{\link[=backtest]{backtest()}}.
}
\description{
Calculates quantile-based scores
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")
ncast <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)

# Run a backtest for the model checking the model fit for two dates:
btest <- backtest(ncast, dates_to_test = c(as.Date("1990-06-11"), as.Date("1990-06-18")))

# Get the rmse with the scoringutils package
if (requireNamespace("scoringutils", quietly = TRUE)){
  metrics = c("wis","dispersion","bias")
  calc_quantile_scores(btest,metrics=metrics)
}

}
