% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{calc_rmse}
\alias{calc_rmse}
\title{calc_rmse}
\usage{
calc_rmse(backtest_summary)
}
\arguments{
\item{backtest_summary}{results of backtest call}
}
\description{
Calculates root mean squared error (rmse)
}
