% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_date_columns}
\alias{check_date_columns}
\title{Check the onset and report dates}
\usage{
check_date_columns(.disease_data, onset_date, report_date)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}
}
\description{
Check the onset and report dates
}
\keyword{internal}
