% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_proportion_reported}
\alias{check_proportion_reported}
\title{Check the \code{proportion_reported} argument}
\usage{
check_proportion_reported(proportion_reported)
}
\arguments{
\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
to less than 1.}
}
\value{
(invisibly) \code{TRUE} if the proportion is valid
}
\description{
Check the \code{proportion_reported} argument
}
\keyword{internal}
