% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{check_same_columns}
\alias{check_same_columns}
\title{Check same columns}
\usage{
check_same_columns(df_list)
}
\arguments{
\item{df_list}{list of dataframes}
}
\description{
Checks that the given list of dataframes contains the same columns
}
\keyword{internal}
