% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_temporal_effect}
\alias{check_temporal_effect}
\title{Check the \code{temporal_effect} argument among the options}
\usage{
check_temporal_effect(t_effect)
}
\arguments{
\item{t_effect}{Either \code{"auto"} to infer the temporal effect or a \code{temporal_effect} object
constructed with the \code{\link[=temporal_effect]{temporal_effect()}} function.}
}
\value{
(invisibly) \code{TRUE} if the temporal effect is valid
}
\description{
Check the \code{temporal_effect} argument among the options
}
\keyword{internal}
