% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_units}
\alias{check_units}
\title{Check the `units` argument among the options}
\usage{
check_units(units)
}
\arguments{
\item{units}{Time scale of reporting. Options: "1 day", "1 week".}
}
\value{
(invisibly) `TRUE` if the units are valid
}
\description{
Check the `units` argument among the options
}
\keyword{internal}
