% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{control_default}
\alias{control_default}
\title{Return the default `STAN` control parameters for nowcasting}
\usage{
control_default()
}
\value{
A list with the default control parameters for running the sampling algorithm.
}
\description{
The function estimates the default control parameters for running [rstan::sampling()].
}
\keyword{internal}
