% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dates.R
\name{datecol}
\alias{datecol}
\title{Get the date column if column is in the data.}
\usage{
datecol(.date_data, .colname)
}
\arguments{
\item{.date_data}{A data frame}

\item{.colname}{Column to check whether it is or isn't in date data}
}
\value{
The column \code{.colname} if it is a column; or an empty vector.
}
\description{
Get the date column if column is in the data.
}
\keyword{internal}
