% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_continuous_priors}
\alias{default_continuous_priors}
\title{Get nowcast priors for the continuous case}
\usage{
default_continuous_priors(...)
}
\value{
A list of the default priors for the nowcasting model when the distribution is continuous
}
\description{
Function to return the priors when the data is assumed to follow a continuous distribution
}
\keyword{internal}
