% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_discrete_priors}
\alias{default_discrete_priors}
\title{Get nowcast priors for the discrete case}
\usage{
default_discrete_priors(...)
}
\arguments{
\item{...}{Name of the parameters for the priors to pass to [set_priors()].}
}
\value{
A list of the default priors for the nowcasting model when the distribution is discrete.
}
\description{
Function to return the priors when the data is assumed to follow a discrete distribution
}
\keyword{internal}
