% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{default_y_link}
\alias{default_y_link}
\title{Default y link}
\usage{
default_y_link(dist)
}
\arguments{
\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}
}
\value{
The name of the default link for a distribution
}
\description{
Return the default links for the model
}
\details{
The functions return the corresponding default links for the data model
}
\keyword{internal}
