% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denguedat}
\alias{denguedat}
\title{denguedat: Dengue fever individual-level reporting data from Puerto Rico}
\format{
A data frame.
}
\usage{
data(denguedat)
}
\description{
Surveillance data from CDC Division of Vector-Borne Diseases.
1990-2010 case reporting data included.
}
\details{
Each row represents a case with the columns indicating the following:
\itemize{
\item \code{onset_week}:  the week of symptom onset.
\item \code{report_week}: the week of case report.
\item \code{gender}: the gender of the infected individual (randomly assigned with 0.5:0.5 probability
of "Male"/"Female").
}
}
\note{
Data originally from the \code{NobBS} package. While \code{onset_week} and \code{report_week} correspond to
actual observed data the \code{gender} was constructed exclusively for the examples of \code{NobBS}. Its
a synthetic (simulated) variable and does not correspond to any reality.
}
\examples{
data(denguedat)
}
\references{
MCGOUGH, Sarah F., et al. Nowcasting by Bayesian Smoothing: A flexible, generalizable model for
real-time epidemic tracking. PLoS computational biology, 2020, vol. 16, no 4, p. e1007735.
}
\keyword{dengue}
