% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasenowcasting-package.R
\docType{package}
\name{diseasenowcasting-package}
\alias{diseasenowcasting}
\alias{diseasenowcasting-package}
\title{diseasenowcasting: Nowcast Time Series of Epidemiological Cases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements Bayesian nowcasting models for incident disease cases using a censored point-process approach. The reporting delay is modelled directly as a stochastic process, jointly with the epidemic dynamics, via a censored likelihood. Inference uses RTMB (CppAD autodiff + Laplace approximation). Supports multiple epidemic models (HSGP, AR1, SIR), multiple delay families (LogNormal, GenGamma, Dirichlet), stratified data, extreme-delay detection, and backtesting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rodrigozepeda.github.io/diseasenowcasting/}
  \item \url{https://github.com/RodrigoZepeda/diseasenowcasting}
  \item Report bugs at \url{https://github.com/RodrigoZepeda/diseasenowcasting/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com} (\href{https://orcid.org/0000-0003-4471-5270}{ORCID})

Authors:
\itemize{
  \item Rami Yaari (\href{https://orcid.org/0000-0002-8808-8937}{ORCID})
  \item Matteo Perini (\href{https://orcid.org/0000-0002-9465-6216}{ORCID})
}

Other contributors:
\itemize{
  \item Tal Robin (\href{https://orcid.org/0000-0002-3128-2499}{ORCID}) [contributor]
  \item Teresa Yamana (\href{https://orcid.org/0000-0001-8349-3151}{ORCID}) [contributor]
  \item Jeffrey Shaman [contributor]
  \item Columbia University in the City of New York [copyright holder, funder]
}

}
\keyword{internal}
