% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{generate_nowcast_dates}
\alias{generate_nowcast_dates}
\title{generate_nowcast_dates}
\usage{
generate_nowcast_dates(start_date, end_date, units, stride)
}
\arguments{
\item{start_date}{start_date}

\item{end_date}{end_date}

\item{stride}{the number of time steps between two consecutive nowcasts}
}
\description{
Generates dates for the backtesting procedure given the start_date, end_date and stride
}
