% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_link_number}
\alias{get_link_number}
\title{Function for sets the link required}
\usage{
get_link_number(link)
}
\value{
A number for the distribution code for Stan's `data`.
}
\description{
Function for sets the link required
}
\keyword{internal}
