% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{get_prior_code_stan}
\alias{get_prior_code_stan}
\title{Function for mapping the priors to numbers}
\usage{
get_prior_code_stan(prior_name)
}
\arguments{
\item{prior_name}{(character) The name of the prior distribution. Can be one of
the following: `standard_normal`, `normal`, `student_t`, `cauchy`, `exponential`,
`gamma`, `inverse_gamma`, `lognormal`, `weibull`, `frechet`,
`double_exponential`, `rayleigh`, `loglogistic`, `gumbel`.}
}
\description{
Takes a prior from one of the listed in `priors.stan` and
returns the number that codes it
}
\keyword{internal}
