% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_priors_from_init}
\alias{get_priors_from_init}
\title{Function for setting the initial values from priors for simulation}
\usage{
get_priors_from_init(priors, num_strata, num_delays, num_steps, dist)
}
\arguments{
\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}

\item{num_strata}{Integer. Number of strata for the population. Default is 2.}

\item{num_delays}{Integer. Number of delay strata. Default is 8.}

\item{num_steps}{Integer. Number of time steps to simulate. Default is 10.}

\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}
}
\value{
An init function that generates initial values based upon the priors to send to
any `rstan` sampling or optimization algorithm.
}
\description{
Function for setting the initial values from priors for simulation
}
\keyword{internal}
