% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{get_quantile_metrics}
\alias{get_quantile_metrics}
\title{get_quantile_metrics}
\usage{
get_quantile_metrics()
}
\description{
Returns vector of supported quantile metrics based on \code{\link[scoringutils]{get_metrics.forecast_quantile}} in \code{\link{scoringutils}}.
}
