% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dates.R
\name{has_date}
\alias{has_date}
\title{Check whether the date column is in the date data}
\usage{
has_date(.date_data, .colname)
}
\arguments{
\item{.date_data}{A data frame}

\item{.colname}{Column to check whether it is or isn't in date data}
}
\value{
Either true or false whether \code{.colname} is in \code{.date_data}.
}
\description{
Check whether the date column is in the date data
}
\keyword{internal}
