% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{infer_date}
\alias{infer_date}
\alias{infer_start_date}
\alias{infer_end_date}
\title{Infer the dates for backtesting}
\usage{
infer_start_date(ncast)

infer_end_date(ncast, start_date)
}
\arguments{
\item{ncast}{A \code{nowcaster} object generated by \code{\link[=nowcast]{nowcast()}}}

\item{start_date}{An object of datatype \code{Date} indicating the date at which
to start the historical nowcasting (first 'now' date to nowcast).}
}
\value{
A date from which backtesting is possible to start or end (respectively)
}
\description{
Infers the start (initial) or  end date one needs to consider for backtesting
}
