% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_weekday}
\alias{is_weekday}
\title{Check whether a date is a weekday vs weekend}
\usage{
is_weekday(date)
}
\arguments{
\item{date}{A date object}
}
\description{
Function that checks whether a date object is a weekday or weekend.
}
\examples{
is_weekday(as.Date("2020-04-22"))
is_weekday(as.Date("2020-04-19"))

}
\references{
From https://stackoverflow.com/a/60346779/5067372
}
