% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpoxdat}
\alias{mpoxdat}
\title{mpoxdat: Mpox reporting data from the 2022 New York City outbreak}
\format{
A data frame.
}
\usage{
data(mpoxdat)
}
\description{
Surveillance line list data provided by the New York City (NYC) Health Department
at https://github.com/nychealth/mpox_nowcast_eval, to accompany a nowcasting
performance evaluation (doi: 10.2196/56495). Patients with a confirmed or probable mpox
diagnosis or illness onset from July 8 through September 30, 2022 were included. The original
dataset was aggregated and pre-processed as described in the note below.
}
\details{
This is line-list data with each row representing case counts. The columns are as follows:
\itemize{
\item \code{dx_date}: is the specimen collection date of the first positive mpox laboratory result,
\item \code{dx_report_date}: is the date the report of first positive mpox laboratory result was received by the NYC Health Department,
\item \code{n}: the case count of individuals within those dates.
\item \code{race}: the race corresponding to those cases. Race was randomly assigned with probabilities
"Non-Hispanic White" = 0.309, "Hispanic" = 0.283, "Black" = 0.202, "Asian" = 0.156, and "Other" = 0.05
which follow what has been reported for the US Census.
}
}
\note{
While \code{dx_date}, \code{dx_report_date} and \code{n} correspond to actual observed data the \code{race} was
constructed exclusively for the examples of this package. Its a synthetic (simulated) variable
and does not correspond to any reality.
}
\examples{
data(mpoxdat)
}
\references{
ROHRER, Rebecca, et al. Nowcasting to Monitor Real-Time Mpox Trends During the 2022
Outbreak in New York City: Evaluation Using Reportable Disease Data Stratified by Race
or Ethnicity. Online Journal of Public Health Informatics, 2025, vol. 17, no 1, p. e56495.
}
\keyword{mpox}
