% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.backtest_metrics}
\alias{plot.backtest_metrics}
\title{plot.backtest_metrics}
\usage{
\method{plot}{backtest_metrics}(x, ..., metric = "wis", horizons = 0, datesbrakes = NULL)
}
\arguments{
\item{x}{results of calls to \code{\link[=backtest_metrics]{backtest_metrics()}}}

\item{...}{further arguments passed to the generic plot function.}

\item{metric}{the metric to display. Specify only one. Default = "wis"}

\item{horizons}{vector of horizons for which the metrics be calculated. Default is to calculate metrics only for horizon 0 (the nowcast).}

\item{datesbrakes}{A string giving the distance between x-axis breaks if not specified, one label
per bar. Other valid examples are: "2 weeks", or "10 years", "sec", "min", "hour", "day", "week",
"month", "year", optionally followed by "s".}
}
\description{
Create a plot to compare the metrics calculated with \code{\link[=backtest_metrics]{backtest_metrics()}}
}
\examples{
#' These examples require the `scoringutils` package
if (requireNamespace("scoringutils", quietly = TRUE)) {
  # Load the data
  data(denguedat)
  # In this example, we will test two models
  now    <- as.Date("1990-10-01")
  ncast1 <- nowcast(denguedat, "onset_week", "report_week", now = now,
                     method = "optimization", seed = 2495624, iter = 10)
  ncast2 <- nowcast(denguedat, "onset_week", "report_week", now = now,
                    method = "optimization", seed = 2495624, iter = 10,
                    dist = "Normal")
  # Run a backtest for each of the models
  btest1 <- backtest(ncast1, dates_to_test = as.Date("1990-06-11"),
                     model_name = "Classic")
  btest2 <- backtest(ncast2, dates_to_test = as.Date("1990-06-11"),
                      model_name = "Normal")
  # Compare the models to select the best model
  comparison <- backtest_metrics(btest1, btest2)
  # plot the comparison
  # plot(comparison)                  ADJ AFTER FIX
  # specify metric and horizons
  # plot(comparison, metric = "mse", horizons = c(-1, 0))
}
}
