% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot.nowcaster}
\alias{plot.nowcaster}
\title{Plot your nowcast}
\usage{
\method{plot}{nowcaster}(
  nowcast_output,
  color = "deepskyblue4",
  datesbrakes = NULL,
  casesbrakes = 10,
  rowsfacet = NULL,
  colsfacet = NULL,
  quantiles = c(0.025, 0.975)
)
}
\arguments{
\item{nowcast_output}{The output of the diseasenowcasting::nowcast() function}

\item{color}{A string indicating the color for the barplots,
Works with the default R colors of grDevices::colors(). Hex color codes works as well. Best results with dark colors}

\item{datesbrakes}{A string giving the distance between x-axis breaks
if \code{NULL}, one label per bar
Other valid examples are:
"2 weeks", or "10 years", 'sec', 'min', 'hour', 'day', 'week', 'month', 'year',
optionally followed by 's'.}

\item{casesbrakes}{Desired number of y-axis breaks. You may get slightly more or fewer breaks that requested.}

\item{rowsfacet}{Number of rows for arranging facets when using ggplot2::facet_wrap().
It allows manual control over the layout of multiple strata in the plot.
If \code{NULL}, the number of rows is automatically determined by ggplot2.}

\item{colsfacet}{Number of columns for arranging facets when using ggplot2::facet_wrap().
It allows manual control over the layout of multiple strata in the plot.
If \code{NULL}, the number of columns is automatically determined by ggplot2.}

\item{quantiles}{a vector of two values to specify the quantiles for the error bars.}
}
\value{
ggplot2 barplots to show real cases and the predictions of the function diseasenowcasting::nowcast().
It creates a subplot for each strata
}
\description{
Create barplots to show real cases and the predictions of the function diseasenowcasting::nowcast()
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")

# perform nowcasting
predictions <- nowcast(denguedat, "onset_week", "report_week",strata = "gender",
                      now = now, method = "optimization",
                      seed = 2495624, iter = 10
                      )


# plot results requires ggplot2 package
if (!requireNamespace("ggplot2", quietly = TRUE)){
  plot(predictions)

  # other plotting options
  plot(predictions, rowsfacet = 2,
       datesbrakes = "2 weeks", casesbrakes = 15,
       color = "firebrick2")
}

}
