% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_nowcast}
\alias{plot_nowcast}
\title{Barplots for the diseasenowcasting::nowcast() output}
\usage{
plot_nowcast(
  nowcast_output,
  maincolor = "deepskyblue4",
  datesbrakes = NULL,
  casesbrakes = 10,
  rowsfacet = NULL,
  colsfacet = NULL
)
}
\arguments{
\item{nowcast_output}{The output of the diseasenowcasting::nowcast() function}

\item{maincolor}{A string indicating the color for the barplots,
Works with the default R colors of grDevices::colors(). Hex color codes works as well.}

\item{datesbrakes}{A string giving the distance between x-axis breaks
if `NULL`, one label per bar
Other valid examples are:
"2 weeks", or "10 years", 'sec', 'min', 'hour', 'day', 'week', 'month', 'year',
optionally followed by 's'.}

\item{casesbrakes}{Desired number of y-axis breaks. You may get slightly more or fewer breaks that requested.}

\item{rowsfacet}{Number of rows for arranging facets when using ggplot2::facet_wrap().
These parameters allow manual control over the layout of multiple strata in the plot.
If `NULL`, the number of rows is automatically determined by ggplot2.

#' @param colsfacet Number of columns for arranging facets when using ggplot2::facet_wrap().
These parameters allow manual control over the layout of multiple strata in the plot.
If `NULL`, the number of columns is automatically determined by ggplot2.}
}
\value{
ggplot2 barplots to show real cases and the predictions of the function diseasenowcasting::nowcast()
it creates a subplot for each strata
}
\description{
Create barplots to show real cases and the predictions of the function diseasenowcasting::nowcast()
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")

# perform nowcasting
predictions <- nowcast(denguedat, "onset_week", "report_week",strata = "gender",
                      now = now, method = "optimization",
                      seed = 2495624, iter = 10
                      )


# plot results
plot_nowcast(predictions)

# other plotting options
plot_nowcast(predictions, rowsfacet = 2,
             datesbrakes = "2 weeks", casesbrakes = 15,
             maincolor = "firebrick2"
             )


}
