% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dates.R
\name{preprocess_dates}
\alias{preprocess_dates}
\title{Preprocess the dates to create a covariate matrix}
\usage{
preprocess_dates(.disease_data, date, temporal_effects)
}
\arguments{
\item{.disease_data}{A \code{data.frame} of cases as exported by \code{\link[=preprocess_for_nowcast]{preprocess_for_nowcast()}}}

\item{date}{The name of the date column in \code{.disease_data} to be used for the temporal effects}

\item{temporal_effects}{A \code{temporal_effects} object codifying the temporal effects to be used.}
}
\value{
A \code{data.frame} containing all of the case info
}
\description{
Takes a \code{data.frame} as generated by \code{\link[=preprocess_for_nowcast]{preprocess_for_nowcast()}} and generates a new \code{data.frame}
where the time covariates are codified for the model
}
\examples{
#Get disease
disease_data <- preprocess_for_nowcast(denguedat, "onset_week", "report_week",
    units = "weeks", now = as.Date("1990-03-05"))
date_data    <- preprocess_dates(disease_data, date = "onset_week",
  temporal_effects = temporal_effects(week_of_year = TRUE))

}
