% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strata.R
\name{preprocess_strata}
\alias{preprocess_strata}
\title{Get strata}
\usage{
preprocess_strata(.disease_data, strata)
}
\arguments{
\item{.disease_data}{A data frame for which to collapse the strata}

\item{strata}{A vector with the names of the strata}
}
\value{
A list with three entries:
}
\description{
Function to preprocess the strata from .disease_data
}
\keyword{internal}
