% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{print.nowcaster}
\alias{print.nowcaster}
\title{Print a nowcast object}
\usage{
\method{print}{nowcaster}(x, ...)
}
\arguments{
\item{x}{A \code{nowcaster} object generated by \code{\link[=nowcast]{nowcast()}}}
}
\description{
Generic to print a \code{nowcaster} object
}
\examples{
#' # Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now   <- as.Date("1990-10-01")
ncast <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)
print(ncast)

}
