% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_effects.R
\name{print.temporal_effect}
\alias{print.temporal_effect}
\title{Print temporal effects}
\usage{
\method{print}{temporal_effect}(x, ...)
}
\arguments{
\item{x}{A \code{temporal_effect} object created with \code{\link[=temporal_effects]{temporal_effects()}}}

\item{...}{Additional arguments to pass to \code{print_temporal_effect_internal} (currently don't do anything)}
}
\description{
Print function extension for a \code{temporal_effect} object created with \code{\link[=temporal_effects]{temporal_effects()}}.
Prints all of the temporal characteristics for a model
}
\examples{
print(temporal_effects(day_of_week = TRUE, week_of_year = TRUE))
print(temporal_effects(day_of_week = FALSE, week_of_year = FALSE))

}
