% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_effects.R
\name{print_temporal_effect_internal}
\alias{print_temporal_effect_internal}
\title{CLI for printing temporal effects}
\usage{
print_temporal_effect_internal(x, ...)
}
\arguments{
\item{x}{A \code{temporal_effect} object created with \code{\link[=temporal_effects]{temporal_effects()}}}

\item{...}{Additional arguments to pass to print.}
}
\description{
Print function for printing the \code{temporal_effect}.
}
\examples{
print(temporal_effects(day_of_week = TRUE, week_of_year = TRUE))
print(temporal_effects(day_of_week = FALSE, week_of_year = FALSE))

}
\keyword{internal}
