% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_prior_class.R
\name{priors}
\alias{priors}
\alias{std_normal_prior}
\alias{normal_prior}
\alias{cauchy_prior}
\alias{student_t_prior}
\alias{double_exponential_prior}
\alias{flat_prior}
\alias{positive_flat_prior}
\alias{half_std_normal_prior}
\alias{half_normal_prior}
\alias{half_cauchy_prior}
\alias{half_student_t_prior}
\alias{half_double_exponential_prior}
\alias{gamma_prior}
\alias{weibull_prior}
\alias{inv_gamma_prior}
\alias{lognormal_prior}
\alias{chi_square_prior}
\alias{exponential_prior}
\alias{logistic_prior}
\alias{beta_prior}
\title{Priors for model parameters}
\usage{
std_normal_prior()

normal_prior(mean, sd)

cauchy_prior(location, scale)

student_t_prior(df, location, scale)

double_exponential_prior(mean, sd)

flat_prior()

positive_flat_prior()

half_std_normal_prior()

half_normal_prior(location, scale)

half_cauchy_prior(location, scale)

half_student_t_prior(df, scale)

half_double_exponential_prior(mean, sd)

gamma_prior(shape, rate)

weibull_prior(shape, scale)

inv_gamma_prior(shape, scale)

lognormal_prior(meanlog, sdlog)

chi_square_prior(df)

exponential_prior(rate)

logistic_prior(location, scale)

beta_prior(alpha, beta)
}
\arguments{
\item{mean, location}{Location parameter.}

\item{sd, scale}{Scale parameter (must be > 0).}

\item{df}{Degrees of freedom (must be > 0).}

\item{shape}{Shape parameter (must be > 0).}

\item{rate}{Rate parameter (must be > 0).}

\item{meanlog}{Mean on the log scale (lognormal).}

\item{sdlog}{SD on the log scale (lognormal, must be > 0).}

\item{alpha, beta}{Beta distribution shape parameters (must be > 0).}
}
\value{
A \code{prior_class} object.
}
\description{
Specify a prior distribution for any estimated parameter.  The result is a
\code{prior_class} object that can be passed to any parameter slot in
\code{\link[=poisson_likelihood]{poisson_likelihood()}}, \code{\link[=nb_likelihood]{nb_likelihood()}}, the delay process constructors,
or the epidemic process constructors.
}
\examples{
normal_prior(0, 1)
gamma_prior(2, 0.1)
exponential_prior(1)
flat_prior()

}
