% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_to_numeric}
\alias{priors_to_numeric}
\title{Function for setting the priors to numeric}
\usage{
priors_to_numeric(.disease_data, priors)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}
}
\value{
The same list of priors but with the numeric codes RStan requires for the distributions
}
\description{
Function for setting the priors to numeric
}
\keyword{internal}
