% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{random_priors}
\alias{random_priors}
\title{Function to randomly set the priors}
\usage{
random_priors(...)
}
\arguments{
\item{...}{Any parameter used in \code{\link[=set_priors]{set_priors()}} that will remain fixed (not random).}
}
\description{
Assign random priors. This is mostly used with the \code{\link[=simulate_process]{simulate_process()}} function.
}
\examples{
#Get random priors
random_priors()

#Get random priors except for mu_0_param_1 which is fixed at 1 and nu_0_param_2 fixed at 0.5
random_priors(mu_0_param_1 = 1, nu_0_param_2 = 0.5)

}
