% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_process_for_testing}
\alias{simulate_process_for_testing}
\title{Simulate a process}
\usage{
simulate_process_for_testing(
  num_steps = 10,
  num_delays = 8,
  num_strata = 2,
  dist = c("NegativeBinomial", "Poisson"),
  mu_degree = 1,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_sd_prior = "normal",
  nu_sd_prior = "normal",
  mu_sd_param_1 = 0,
  mu_sd_param_2 = 0.1,
  nu_sd_param_1 = 0,
  nu_sd_param_2 = 0.1,
  mu_0_mean_param_1 = log(100),
  mu_0_mean_param_2 = 0.01,
  mu_0_sd_param_1 = 0.01,
  mu_0_sd_param_2 = 0.01,
  nu_0_mean_param_1 = 0,
  nu_0_mean_param_2 = 0.01,
  nu_0_sd_param_1 = 0,
  nu_0_sd_param_2 = 0.01,
  mu_0_mean_hyperprior = "normal",
  nu_0_mean_hyperprior = "normal",
  mu_0_sd_hyperprior = "normal",
  nu_0_sd_hyperprior = "normal",
  r_prior = "normal",
  r_param_1 = 0,
  r_param_2 = 1
)
}
\arguments{
\item{num_steps}{Integer. Number of time steps to simulate. Default is 10.}

\item{num_delays}{Integer. Number of delay strata. Default is 8.}

\item{num_strata}{Integer. Number of strata for the population. Default is 2.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{mu_degree}{Integer. Degree of the epidemic trend. Default is 1.}

\item{nu_degree}{Integer. Degree of the delay trend. Default is 1.}

\item{mu_is_constant}{Logical. Indicates whether the epidemic trend is constant. Default is FALSE.}

\item{nu_is_constant}{Logical. Indicates whether the delay trend is constant. Default is TRUE.}

\item{mu_sd_prior}{Character. Prior for the epidemic trend error. Default is "normal".}

\item{nu_sd_prior}{Character. Prior for the delay trend error. Default is "normal".}

\item{mu_sd_param_1}{Numeric. First parameter for the epidemic trend error. Default is 0.0.}

\item{mu_sd_param_2}{Numeric. Second parameter for the epidemic trend error. Default is 1.0.}

\item{nu_sd_param_1}{Numeric. First parameter for the delay trend error. Default is 0.0.}

\item{nu_sd_param_2}{Numeric. Second parameter for the delay trend error. Default is 1.0.}

\item{mu_0_mean_param_1}{Character. Prior for the initial epidemic trend mean's first parameter.}

\item{mu_0_mean_param_2}{Character. Prior for the initial epidemic trend mean's second parameter.}

\item{mu_0_sd_param_1}{Character. Prior for the initial epidemic trend standard deviation's first parameter.}

\item{mu_0_sd_param_2}{Character. Prior for the initial epidemic trend standard deviation's second parameter.}

\item{nu_0_mean_param_1}{Character. Prior for the initial delay trend mean's first parameter.}

\item{nu_0_mean_param_2}{Character. Prior for the initial delay trend mean's second parameter.}

\item{nu_0_sd_param_1}{Character. Prior for the initial delay trend standard deviation's first parameter.}

\item{nu_0_sd_param_2}{Character. Prior for the initial delay trend standard deviation's second parameter.}

\item{mu_0_mean_hyperprior}{Prior distribution for the mean of the epidemic trend. Default is normal.}

\item{nu_0_mean_hyperprior}{Prior distribution for the mean of the delay trend. Default is normal.}

\item{mu_0_sd_hyperprior}{Prior distribution for the standard deviation of the epidemic trend. Default is normal (truncated).}

\item{nu_0_sd_hyperprior}{Prior distribution for the standard deviation of the delay trend. Default is normal (truncated).}

\item{r_prior}{Character. Prior for the negative binomial precision parameter. Default is "normal".}

\item{r_param_1}{Numeric. First parameter for the dispersion prior if negative binomial. Default is 0.0.}

\item{r_param_2}{Numeric. Second parameter for the dispersion prior if negative binomial. Default is 1.0.}
}
\value{
A tibble with simulated state-space process results, including the onset and
report dates, strata, delays, and the number of observed cases per time step.
}
\description{
This function simulates a state-space process for testing purposes in a model with
multiple delays and strata. It estimates the process over a specified number of steps,
 with options for trend degrees, priors, and error distributions.
}
\examples{
simulate_process_for_testing(num_steps = 20, num_delays = 5, num_strata = 3)
}
