% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_function_tester.R
\name{stan_function_tester}
\alias{stan_function_tester}
\title{Stan function tester}
\usage{
stan_function_tester()
}
\value{
All of the stan functions in the package inside an environment
}
\description{
Function that loads all of the stan functions developed for the package.
Primarily for testing purposes
}
\keyword{internal}
