% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.backtest_metrics}
\alias{summary.backtest_metrics}
\title{Makes a summary of your \code{\link[=backtest_metrics]{backtest_metrics()}} object}
\usage{
\method{summary}{backtest_metrics}(object, ..., metrics = "all")
}
\arguments{
\item{object}{A backtest_metrics object generated by \code{\link[=backtest_metrics]{backtest_metrics()}}}

\item{...}{further arguments passed to the generic \code{\link[=summary]{summary()}} function.}

\item{metrics}{selects the metrics to display.
\code{metrics = "all"} summarizes all available metrics generated in \code{\link[=backtest_metrics]{backtest_metrics()}}}
}
\value{
A summary tibble of the \code{\link[=backtest_metrics]{backtest_metrics()}} results,
it specifies: average, standard deviation, standard error, number of "now" dates
it also returns the metrics across all the horizons provided ("all" in column "horizon")
}
\description{
Makes a tidy summary dataframe of the results of the function \code{\link[=backtest_metrics]{backtest_metrics()}}
}
\examples{
# These examples require the `scoringutils` package
if (requireNamespace("scoringutils", quietly = TRUE)) {
  # Load the data
  data(denguedat)
  # In this example, we will test two models
  now    <- as.Date("1990-10-01")
  ncast1 <- nowcast(denguedat, "onset_week", "report_week", now = now,
                     method = "optimization", seed = 2495624, iter = 10)
  ncast2 <- nowcast(denguedat, "onset_week", "report_week", now = now,
                    method = "optimization", seed = 2495624, iter = 10,
                    dist = "Normal")
  # Run a backtest for each of the models
  btest1 <- backtest(ncast1, dates_to_test = as.Date("1990-06-11"),
                     model_name = "Classic")
  btest2 <- backtest(ncast2, dates_to_test = as.Date("1990-06-11"),
                      model_name = "Normal")
  # Compare the models to select the best model
  comparison <- backtest_metrics(btest1, btest2)

  # Create a summary
  summary(comparison)

  # Create a summary with more metrics:
  summary(comparison, metrics = c("mae", "rmse", "ape", "wis"))
  }
}
