% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{summary.nowcaster}
\alias{summary.nowcaster}
\title{Makes a summary of your nowcast}
\usage{
\method{summary}{nowcaster}(nowcast_output, quantiles = c(0.025, 0.975))
}
\arguments{
\item{nowcast_output}{the output of the diseasenowcasting::nowcast() function}

\item{quantiles}{a vector to specify the quantiles to display.}
}
\value{
A summary tibble of the nowcast results,
it specifies: onset time, Strata, Mean value, standard deviation, quantiles
}
\description{
Makes a tidy summary dataframe of the results of the function diseasenowcasting::nowcast()
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")

# perform nowcasting
predictions <- nowcast(denguedat, "onset_week", "report_week",
  now = now, method = "optimization",
  seed = 2495624, iter = 10
)

# create summary dataframe
summary(predictions)

}
