% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{summary_nowcast}
\alias{summary_nowcast}
\title{Makes a summary of nowcast() output}
\usage{
summary_nowcast(nowcast_output)
}
\arguments{
\item{nowcast_output}{the output of the nowcast() function}
}
\description{
Makes a tidy summary dataframe of the results of the function nowcast()
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")

# perform nowcasting
predictions <- nowcast(denguedat, "onset_week", "report_week", now = now, method = "optimization",
seed = 2495624, iter = 10)

# create summary dataframe
summary_nowcast(predictions)


}
