% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_effects.R
\name{temporal_effects}
\alias{temporal_effects}
\title{Temporal effects}
\usage{
temporal_effects(
  day_of_week = FALSE,
  weekend = FALSE,
  day_of_month = FALSE,
  month_of_year = FALSE,
  week_of_year = FALSE,
  holidays = NULL
)
}
\arguments{
\item{day_of_week}{Boolean. Whether to include an effect for each of the seven days of the week.}

\item{weekend}{Boolean. Whether to include an effect for the weekend vs the weekday.}

\item{day_of_month}{Boolean. Whether to include an effect for the day of the month (i.e. 1 to 31st)}

\item{month_of_year}{Boolean. Whether to include an effect for the month.}

\item{week_of_year}{Boolean. Whether to include an effect for the epidemiological week.}

\item{holidays}{Either \code{NULL} or an \code{\link[almanac:rcalendar]{almanac::rcalendar()}} specifying how to calculate holidays.}
}
\description{
Matrix for constructing temporal effects into the \code{\link[=nowcast]{nowcast()}}.
}
\section{US holidays}{


US Federal holidays can be passed by just passing the \code{\link[almanac:cal_us_federal]{almanac::cal_us_federal()}} function
into temporal effects.

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(almanac)
temporal_effects(holidays = cal_us_federal())
}\if{html}{\out{</div>}}
}

\examples{
#Create day of week and week of year class
temporal_effects(day_of_week = TRUE, week_of_year = TRUE)

#Set up a holiday effect on Christmass using the almanac package
if(!requireNamespace("almanac", quietly = TRUE)){
  cal <- rcalendar(hol_christmas())
  temporal_effects(holidays = cal)
}
}
