% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25_autoplot.R
\name{theme_diseasenowcasting}
\alias{theme_diseasenowcasting}
\title{ggplot2 theme matching the diseasenowcasting visual identity}
\usage{
theme_diseasenowcasting(base_size = 11)
}
\arguments{
\item{base_size}{Base font size.}
}
\description{
ggplot2 theme matching the diseasenowcasting visual identity
}
