% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update.nowcaster}
\alias{update.nowcaster}
\title{Update a nowcast object}
\usage{
\method{update}{nowcaster}(
  object,
  new_data,
  now = NULL,
  refresh = 250 * rlang::is_interactive(),
  ...
)
}
\arguments{
\item{object}{A \code{nowcaster} object generated by \code{\link[=nowcast]{nowcast()}}}

\item{new_data}{New \code{.disease_data} to use for fitting \code{\link[=nowcast]{nowcast()}}. Ideally this should include
the previous data + new observations in the same \code{data.frame}.}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perlform the nowcast.}

\item{refresh}{Refresh parameter for \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{...}{Additional arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\description{
Generic to update a \code{nowcaster} object with new data
}
\examples{
#' # Load the data
data(denguedat)

# Run a nowcast using the first 1000 data points
first_1k <- denguedat[1:1000,]
ncast <- nowcast(first_1k, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10, refresh = 0)
ncast

#Update the nowcast now that new information was acquired
first_3k <- denguedat[1:3000,]
update(ncast, new_data = first_3k)

}
