% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{backtest}
\usage{
backtest(
  start_date = NULL,
  end_date = NULL,
  stride = 1,
  retrain = 1,
  probs = c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975),
  min_horizon = 0,
  model_name = NULL,
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson", "Normal", "Student"),
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  refresh = 250 * rlang::is_interactive(),
  control = control_default(),
  method = c("sampling", "variational", "optimization"),
  priors = set_priors(),
  ...
)
}
\arguments{
\item{start_date}{An object of datatype \code{Date} indicating the date at which
to start the historical nowcasting (first 'now' date to nowcast).}

\item{end_date}{An object of datatype \code{Date} indicating the date at which
to end the historical nowcasting (last 'now' date to nowcast).}

\item{stride}{Integer variable indicating the number of time steps between
two consecutive nowcasts. Default 1 means nowcasting for each date between
start and end,}

\item{retrain}{Integer variable indicating the number of iterations for which
to retrain the model. Default 1 means retraining the model for each nowcast.
NOT IMPLEMENTED YET - currently will be retrained at each time step.}

\item{probs}{list of probabilities in [0,1] defining which quantile estimates
will be returned}

\item{min_horizon}{the minimum horizon in [-Inf,0] to keep from each nowcast estimates
(e.g. min_horizon=-5 means estimates of up to previous 5 time steps from now will be kept)}

\item{model_name}{A model name that will be used to identify the results of the
backtest with the given parameters in a subsequent call to [backtest_metrics()]}

\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
 to less than 1.}

\item{refresh}{Refresh parameter for [rstan::sampling()]}

\item{control}{Control parameter for [rstan::sampling()]}

\item{method}{Fitting method either `sampling` (recommended for inference), `variational`
(recommended for testing) or `optimization`. The `sampling` method calls [rstan::sampling()]
while the `variational` calls [rstan::vb()] and `optimization` calls [rstan::optimizing()]}

\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}

\item{...}{Additional arguments to pass to [rstan::sampling()]}
}
\description{
Compute historical nowcasts for given time period
}
