% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{Backtest}
\usage{
backtest(
  ncast,
  start_date = infer_start_date(ncast),
  end_date = infer_end_date(ncast, start_date = start_date),
  stride = 1,
  subsample = NULL,
  dates_to_test = generate_nowcast_dates(start_date, end_date, ncast, stride, subsample =
    subsample),
  retrain = 1,
  quantiles = c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975),
  min_horizon = 0,
  model_name = NULL,
  ...
)
}
\arguments{
\item{ncast}{A \code{nowcaster} object generated by \code{\link[=nowcast]{nowcast()}}}

\item{start_date}{An object of datatype \code{Date} indicating the date at which
to start the historical nowcasting (first 'now' date to nowcast).}

\item{end_date}{An object of datatype \code{Date} indicating the date at which
to end the historical nowcasting (last 'now' date to nowcast).}

\item{stride}{Integer variable indicating the number of time steps between
two consecutive nowcasts. Default 1 means nowcasting for each date between
start and end,}

\item{dates_to_test}{Specific values of moments \code{now} to test the nowcast. If \code{dates_to_test}
is given this overrides \code{start_date}, \code{end_date} and \code{stride}.}

\item{retrain}{Integer variable indicating the number of iterations for which
to retrain the model. Default 1 means retraining the model for each nowcast.
NOT IMPLEMENTED YET - currently will be retrained at each time step.}

\item{quantiles}{list of quantiles in \link{0,1} defining which quantile estimates
will be returned}

\item{min_horizon}{the minimum horizon in \link{-Inf,0} to keep from each nowcast estimates
(e.g. min_horizon=-5 means estimates of up to previous 5 time steps from now will be kept)}

\item{model_name}{A model name that will be used to identify the results of the
backtest with the given parameters in a subsequent call to \code{\link[=backtest_metrics]{backtest_metrics()}}}

\item{...}{Additional arguments to pass to \code{\link[=update.nowcaster]{update.nowcaster()}}}
}
\value{
A tibble with as many rows as \code{dates_to_test} are given. Each row represents a different \code{now}
with the following columns:
\itemize{
\item \code{now}: The date nowcasted
\item \code{true_date}: The column corresponding to the \code{true_date} (has the \code{true_date}'s name).
\item \code{mean}: The mean estimate of the \code{nowcast}.
\item \code{sd}: The standard deviation estimate of the \code{nowcast}.
\item \code{median}: The median estimate of the \code{nowcast}.
\item \verb{x\%}: Quantile columns corresponding to the estimated \code{xth} quantile.
\item \code{Strata_unified}: A column containing the strata modeled.
\item \code{observed}: The value actually obnserved for that date
\item \code{model}: The name of the fitted model
}
}
\description{
Compute historical nowcasts for given time period
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")
ncast <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)

# Run a backtest for the model checking the model fit for two dates:
btest <- backtest(ncast, dates_to_test = c(as.Date("1990-06-11"), as.Date("1990-06-18")))

#The following examples are slow:
\dontrun{
# Run a backtest for the model automatically selecting 3 possible dates at random
btest2 <- backtest(ncast, subsample = 3)

# Run a backtest for the model using all possible dates
btest3 <- backtest(ncast)
}
}
