% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_metrics}
\alias{backtest_metrics}
\title{backtest_metrics}
\usage{
backtest_metrics(backtest_summary, metrics, horizons = c(0))
}
\arguments{
\item{backtest_summary}{Results of call to backtest or a list of results of backtest calls.}

\item{metrics}{list of metrics which should be calculated.
Currently supporting:MAE(mean absolute error), RMSE (root mean squared error),
and WIS (weighted interval score)}

\item{horizons}{list of horizons for which the metrics be calculated.
Default is to calculate metrics only for horizon 0.}
}
\description{
Compute metrics for given backtest results
}
