% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{calc_rmse}
\alias{calc_rmse}
\title{calc_rmse}
\usage{
calc_rmse(backtest_summary)
}
\arguments{
\item{backtest_summary}{results of backtest call}
}
\value{
The Root Mean Squared Error for each of the runs in \code{\link[=backtest]{backtest()}}.
}
\description{
Calculates root mean squared error (rmse)
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")
ncast <- nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)

# Run a backtest for the model checking the model fit for two dates:
btest <- backtest(ncast, dates_to_test = c(as.Date("1990-06-11"), as.Date("1990-06-18")))

# Get the rmse with the scoring utils package
if (requireNamespace("scoringutils", quietly = TRUE)){
  calc_rmse(btest)
}

}
