% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_date_columns}
\alias{check_date_columns}
\title{Check the onset and report dates}
\usage{
check_date_columns(.disease_data, true_date, report_date)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{true_date}{In quotations, the name of the column designating the true date
when the case happend e.g. "onset_week". The datatype of the column should be a \code{Date}.}

\item{report_date}{In quotations, the name of the column designating the date
the case was reported e.g. "report_week". The datatype of the column should be a \code{Date}.}
}
\value{
The `.disease_data` data.frame filtered for reports after onset.
}
\description{
Check the onset and report dates
}
\keyword{internal}
