% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_priors}
\alias{default_priors}
\title{Get nowcast priors for the continuous case}
\usage{
default_priors(dist, ...)
}
\arguments{
\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}

\item{...}{Additional arguments to pass to [rstan::sampling()]}
}
\value{
A list of the default priors for the nowcasting model when the distribution is discrete.
}
\description{
Function to return the priors when the data is assumed to follow a continuous distribution
}
\examples{

#Use inside nowcasting as dollows:
data(denguedat)

# Running a quick nowcast with almost zero iterations
now <- as.Date("1990-10-01")
nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10,
  priors = default_priors("Normal", p = 0, q = 1)) #Change whatever you want from the defaults


}
