% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denguedat.R
\docType{data}
\name{denguedat}
\alias{denguedat}
\title{denguedat: Dengue fever reporting data from Puerto Rico}
\format{
A data frame.
}
\usage{
data(denguedat)
}
\description{
Surveillance data from CDC Division of Vector-Borne Diseases.
1990-2010 case reporting data included. Data originally from the `NobBS`
package.
The first column, `onset_week`, indicates the week of symptom onset.
The second column, `report_week`, indicates the week of case report.
The third column, `gender`, indicates the gender of the infected individual
(randomly assigned with 0.5:0.5 probability of "Male"/"Female").
}
\examples{
data(denguedat)
}
\keyword{dengue}
