% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasenowcasting-package.R
\docType{package}
\name{diseasenowcasting-package}
\alias{diseasenowcasting}
\alias{diseasenowcasting-package}
\title{diseasenowcasting: Nowcasting disease onset cases given report dates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements bayesian state-space models to nowcast incident cases for any disease given report dates while controlling by strata and covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rodrigozepeda.github.io/diseasenowcasting/}
}

}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com} (\href{https://orcid.org/0000-0003-4471-5270}{ORCID})

Authors:
\itemize{
  \item Rami Yaari (\href{https://orcid.org/0000-0002-8808-8937}{ORCID})
  \item Matteo Perini (\href{https://orcid.org/0000-0002-9465-6216}{ORCID})
}

Other contributors:
\itemize{
  \item Tal Robin (\href{https://orcid.org/0000-0002-3128-2499}{ORCID}) [contributor]
  \item Teresa Yamana (\href{https://orcid.org/0000-0001-8349-3151}{ORCID}) [contributor]
  \item Columbia University in the City of New York [copyright holder, funder]
}

}
\keyword{internal}
