% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasenowcasting-package.R
\docType{package}
\name{diseasenowcasting-package}
\alias{diseasenowcasting}
\alias{diseasenowcasting-package}
\title{diseasenowcasting: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
