% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{generate_nowcast_dates}
\alias{generate_nowcast_dates}
\title{Generate nowcast dates}
\usage{
generate_nowcast_dates(start_date, end_date, ncast, stride, subsample = NULL)
}
\arguments{
\item{start_date}{start_date}

\item{end_date}{end_date}

\item{ncast}{A nowcaster object generated with \code{\link[=nowcast]{nowcast()}}}

\item{stride}{the number of time steps between two consecutive nowcasts}

\item{subsample}{If an integer, how much to subsample the dates; if NULL then runs all dates.}
}
\description{
Generates dates for the backtesting procedure given the start_date, end_date and stride
}
\keyword{internal}
