% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_link_number}
\alias{get_link_number}
\title{Function for sets the link required}
\usage{
get_link_number(link)
}
\value{
A number for the distribution code for Stan's `data`.
}
\description{
The following links are implemented:
* identity: g(x) = x
* log: g(x) = log(x)
* softplus: g(x) = k*log(1 + exp(x/k))
* dhyperbolic: g(x) = c*x/2 + sqrt( (c*x/2)^2 + k)
}
\keyword{internal}
