% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{get_prior_code_stan}
\alias{get_prior_code_stan}
\title{Function for mapping the priors to numbers}
\usage{
get_prior_code_stan(prior_name)
}
\arguments{
\item{prior_name}{(character) The name of the prior distribution. Can be one of
the following: \code{jeffreys} (non-informative), \code{standard_normal}, \code{normal}, \code{student_t},
\code{cauchy}, \code{exponential}, \code{gamma}, \code{inverse_gamma}, \code{lognormal}, \code{weibull}, \code{frechet},
\code{double_exponential}, \code{rayleigh}, \code{loglogistic}, \code{gumbel}.}
}
\description{
Takes a prior from one of the listed in \code{priors.stan} and
returns the number that codes it
}
\keyword{internal}
