% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_now}
\alias{infer_now}
\title{Automatically infer which value is `now`.}
\usage{
infer_now(.disease_data, now, onset_date)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}
}
\description{
Function returns the maximum onset date of `.disease_data` if `now = NULL`. Else
it check whether now is possible in the data.
}
\keyword{internal}
