% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpoxdat}
\alias{mpoxdat}
\title{mpoxdat: M-pox line-list reporting data for New York City}
\format{
A data frame.
}
\usage{
data(mpoxdat)
}
\description{
Surveillance data from from the New York City Health Department’s mpox surveillance database.
}
\details{
This is line-list data with each row representing case counts. The columns are as follows:
\itemize{
\item \code{dx_date}: first positive mpox laboratory result
\item \code{dx_report_date}: the first positive mpox laboratory result was received by the NYC Health Department
\item \code{n}: the case count of individuals within those dates.
\item \code{race}: the race corresponding to those cases. Race was randomly assigned with probabilities
"Non-Hispanic White" = 0.309, "Hispanic" = 0.283, "Black" = 0.202, "Asian" = 0.156, and "Other" = 0.05
which follow what has been reported for the US Census.
}
}
\note{
While \code{dx_date}, \code{dx_report_date} and \code{n} correspond to actual observed data the \code{race} was
constructed exclusively for the examples of this package. Its a synthetic (simulated) variable
and does not correspond to any reality.
}
\examples{
data(mpoxdat)
}
\references{
ROHRER, Rebecca, et al. Nowcasting to Monitor Real-Time Mpox Trends During the 2022
Outbreak in New York City: Evaluation Using Reportable Disease Data Stratified by Race
or Ethnicity. Online Journal of Public Health Informatics, 2025, vol. 17, no 1, p. e56495.
}
\keyword{mpox}
