% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting}
\usage{
nowcast(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  now = NULL,
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  refresh = 250 * interactive(),
  control = control_default(),
  mu_degree = 2,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_error_prior = "normal",
  nu_error_prior = "normal",
  mu_param_1 = 0,
  mu_param_2 = 1,
  nu_param_1 = 0,
  nu_param_2 = 1,
  mu_0_prior = "normal",
  nu_0_prior = "normal",
  mu_0_param_1 = log(mean(.disease_data$n, na.rm = T)),
  mu_0_param_2 = 1,
  nu_0_param_1 = 0,
  nu_0_param_2 = 1,
  r_prior = "normal",
  r_param_1 = 0,
  r_param_2 = 1,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
 to less than 1.}

\item{refresh}{Refresh parameter for [rstan::sampling()]}

\item{control}{Control parameter for [rstan::sampling()]}

\item{mu_degree}{Integer. Degree of the epidemic trend. Default is 2.}

\item{nu_degree}{Integer. Degree of the delay trend. Default is 1.}

\item{mu_is_constant}{Logical. Indicates whether the epidemic trend is constant. Default is FALSE.}

\item{nu_is_constant}{Logical. Indicates whether the delay trend is constant. Default is TRUE.}

\item{mu_error_prior}{Character. Prior for the epidemic trend error. Default is "normal".}

\item{nu_error_prior}{Character. Prior for the delay trend error. Default is "normal".}

\item{mu_param_1}{Numeric. First parameter for the epidemic trend error. Default is 0.0.}

\item{mu_param_2}{Numeric. Second parameter for the epidemic trend error. Default is 1.0.}

\item{nu_param_1}{Numeric. First parameter for the delay trend error. Default is 0.0.}

\item{nu_param_2}{Numeric. Second parameter for the delay trend error. Default is 1.0.}

\item{mu_0_prior}{Character. Prior for the initial epidemic trend distribution. Default is "normal".}

\item{nu_0_prior}{Character. Prior for the initial delay trend distribution. Default is "normal".}

\item{mu_0_param_1}{Numeric. First parameter for the initial epidemic trend distribution. Default is 0.0.}

\item{mu_0_param_2}{Numeric. Second parameter for the initial epidemic trend distribution. Default is 1.0.}

\item{nu_0_param_1}{Numeric. First parameter for the initial delay trend distribution. Default is 0.0.}

\item{nu_0_param_2}{Numeric. Second parameter for the initial delay trend distribution. Default is 1.0.}

\item{r_prior}{Character. Prior for the negative binomial precision parameter. Default is "normal".}

\item{r_param_1}{Numeric. First parameter for the dispersion prior if negative binomial. Default is 0.0.}

\item{r_param_2}{Numeric. Second parameter for the dispersion prior if negative binomial. Default is 1.0.}

\item{...}{Additional arguments to pass to [rstan::sampling()]}

\item{is_negative_binomial}{Logical. Indicates whether the data is modeled using a negative binomial distribution. Default is TRUE.}
}
\description{
Function that uses the [rstan::sampling()] engine to generate nowcasts.
}
\examples{
# Load the data
data(denguedat)
now <- as.Date("1990-10-01")

# Run a nowcast with very few iterations
# change to 4 chains and 2000 iter when doing inference
# or just run the command without the `iter = 50`, `chains = 1` thing.
nowcast(denguedat, "onset_week", "report_week", strata = c("gender"), now = now,
          #The following is specified as to run fast for the example;
          iter = 50, chains = 1, seed = 2524)
}
