% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting}
\usage{
nowcast(
  .disease_data,
  onset_date,
  report_date,
  covariates = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  now = NULL,
  units = NULL,
  seed = NULL,
  moving_window = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{moving_window}{Size of moving window for estimation of cases (numeric).
The moving window size should be specified in the same date units as the
reporting data (i.e. specify 7 to indicate 7 days, 7 weeks, etc).
Default: NULL, i.e. takes all historical dates into consideration.}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
 to less than 1.}

\item{...}{Additional arguments to pass to Stan.}
}
\description{
Function that uses the `brms` engine to generate nowcasts.
}
\examples{
# Load the data
data(denguedat)
now <- as.Date("1990-10-01")
nowcast(denguedat, "onset_week", "report_week")
}
