% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting}
\usage{
nowcast(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  now = NULL,
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  refresh = 10 * interactive(),
  control = control_default(),
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
 to less than 1.}

\item{...}{Additional arguments to pass to [rstan::sampling()]}
}
\description{
Function that uses the [rstan::sampling()] engine to generate nowcasts.
}
\examples{
# Load the data
data(denguedat)
now <- as.Date("1990-10-01")

# Run a nowcast with very few iterations
# change to 4 chains and 2000 iter when doing inference
# or just run the command without the `iter = 10`, `chains = 1` thing.
nowcast(denguedat, "onset_week", "report_week",
          iter = 100, chains = 1, seed = 12334)

# You can also run the nowcast stratifying by several variables
nowcast(denguedat, "onset_week", "report_week", strata = c("gender"),
          iter = 100, chains = 1, seed = 12334)
}
