% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcasting}
\usage{
nowcast(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson", "Normal", "Student"),
  link_x = default_x_link(dist),
  link_y = default_y_link(dist),
  now = NULL,
  units = NULL,
  max_delay = Inf,
  prior_only = FALSE,
  proportion_reported = 1,
  normalize_data = (dist \%in\% c("Normal", "Student")),
  refresh = 250 * rlang::is_interactive(),
  control = control_default(),
  method = c("sampling", "variational", "optimization"),
  priors = set_priors(),
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{proportion_reported}{A decimal greater than 0 and less than or equal to
1 representing the proportion of all cases expected to be reported.
Default: 1, e.g. 100 percent of all cases will eventually be reported.
For asymptomatic diseases where not all cases will ever be reported,
or for outbreaks in which severe under-reporting is expected, change this
 to less than 1.}

\item{refresh}{Refresh parameter for [rstan::sampling()]}

\item{control}{Control parameter for [rstan::sampling()]}

\item{method}{Fitting method either `sampling` (recommended for inference), `variational`
(recommended for testing) or `optimization`. The `sampling` method calls [rstan::sampling()] while the
`variational` calls [rstan::vb()] and `optimization` calls [rstan::optimizing()]}

\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}

\item{...}{Additional arguments to pass to [rstan::sampling()]}
}
\description{
Function that uses the [rstan::sampling()] engine to generate nowcasts.
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")
nowcast(denguedat, "onset_week", "report_week", now = now,
  method = "optimization", seed = 2495624, iter = 10)
}
