% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast.rstan}
\alias{nowcast.rstan}
\title{Nowcasting with the `rstan` engine}
\usage{
nowcast.rstan(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  prior_only = FALSE,
  init = 0,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{prior_only}{Boolean variable indicating whether to compute only the prior}

\item{init}{Initial values for [rstan::sampling()]}

\item{...}{Additional arguments to pass to [rstan::sample()]}
}
\description{
Nowcasting with the `rstan` engine
}
\keyword{internal}
