% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast.rstan}
\alias{nowcast.rstan}
\title{Nowcasting with the `rstan` engine}
\usage{
nowcast.rstan(
  .disease_data,
  onset_date,
  report_date,
  strata = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  mu_degree = 2,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  prior_only = FALSE,
  mu_shape_prior = 0,
  mu_rate_prior = 1,
  nu_shape_prior = 0,
  nu_rate_prior = 1,
  dispersion_prior_shape = 0,
  dispersion_prior_rate = 0.001,
  mean_mu_0_prior = log(mean(.disease_data$n, na.rm = T)),
  mean_nu_0_prior = 0,
  sigma_mu_0_prior = 0.01,
  sigma_nu_0_prior = 0.01,
  mu_prior_name = "standard_normal",
  nu_prior_name = "standard_normal",
  r_prior_name = "normal",
  control = control_default(),
  refresh = 250,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{strata}{Character vector of names of the strata included in the data.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{control}{Control parameter for [rstan::sampling()]}

\item{refresh}{Refresh parameter for [rstan::sampling()]}

\item{...}{Additional arguments to pass to [rstan::sample()]}
}
\description{
Nowcasting with the `rstan` engine
}
\keyword{internal}
