% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast.rstan}
\alias{nowcast.rstan}
\title{Nowcasting with the \code{rstan} engine}
\usage{
nowcast.rstan(
  .disease_data,
  .date_epidemic,
  .date_delay,
  true_date,
  report_date,
  num_steps,
  num_delays,
  num_strata,
  dist,
  link_x,
  link_y,
  normalize_data,
  has_cycle,
  autoregresive,
  moving_average,
  prior_only,
  control,
  refresh,
  method,
  priors,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and (optionally) strata}

\item{true_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{num_steps}{Number of time steps to run in the model}

\item{num_delays}{Number of delays to consider in the model}

\item{num_strata}{Number of strata in the model}

\item{dist}{Distribution. Either "NegativeBinomial", "Poisson", "Normal", or "Student"}

\item{link_x}{Link function for the epidemic process (see section On links).}

\item{link_y}{Link function for the data (see section On links).}

\item{normalize_data}{Whether the data \code{y} should be normalized (substracted its mean and divided
by standard deviation) for fitting. This option is only valid if using a continuous model
(Normal or Student).}

\item{has_cycle}{Boolean. Whether include a cycle component in the model.}

\item{autoregresive}{An \code{\link[=AR]{AR()}} object with the autoregresive components for the epidemic
and delay processes.}

\item{moving_average}{A \code{\link[=MA]{MA()}} object with the moving average component for the epidemic
and delay processes.}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{control}{Control parameter for \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{refresh}{Refresh parameter for \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{method}{Fitting method either \code{sampling} (recommended for inference), \code{variational}
(recommended for testing) or \code{optimization}. The \code{sampling} method calls \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}
while the \code{variational} calls \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} and \code{optimization} calls \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}}

\item{priors}{A list of all of the nowcast priors. You can use \code{\link[=set_priors]{set_priors()}} to change
the priors of the function (see details)}

\item{...}{Additional arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\description{
Nowcasting with the \code{rstan} engine
}
\keyword{internal}
