% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast.rstan}
\alias{nowcast.rstan}
\title{Nowcasting with specific engines}
\usage{
nowcast.rstan(
  .disease_data,
  onset_date,
  covariates = NULL,
  dist = c("NegativeBinomial", "Poisson"),
  prior_only = FALSE,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{...}{Additional arguments to pass to [rstan::sample()]}
}
\description{
Nowcasting with specific engines
}
\examples{
# Load the data
data(denguedat)
now <- as.Date("1990-10-01")
nowcast(denguedat, onset_dat)
}
