% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast.rstan}
\alias{nowcast.rstan}
\title{Nowcasting with the `rstan` engine}
\usage{
nowcast.rstan(
  .disease_data,
  true_date,
  report_date,
  num_steps,
  num_delays,
  num_strata,
  dist,
  prior_only,
  control,
  refresh,
  method,
  priors,
  ...
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{true_date}{In quotations, the name of the column designating the true date
when the case happend e.g. "onset_week". The datatype of the column should be a \code{Date}.}

\item{report_date}{In quotations, the name of the column designating the date
the case was reported e.g. "report_week". The datatype of the column should be a \code{Date}.}

\item{num_steps}{Number of time steps to run in the model}

\item{num_delays}{Number of delays to consider in the model}

\item{num_strata}{Number of strata in the model}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{prior_only}{Boolean variable indicating whether to compute only the prior distribution}

\item{control}{Control parameter for [rstan::sampling()]}

\item{refresh}{Refresh parameter for [rstan::sampling()]}

\item{method}{Fitting method either `sampling` (recommended for inference), `variational`
(recommended for testing) or `optimization`. The `sampling` method calls [rstan::sampling()] while the
`variational` calls [rstan::vb()] and `optimization` calls [rstan::optimizing()]}

\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}

\item{...}{Additional arguments to pass to [rstan::sampling()]}
}
\description{
Nowcasting with the `rstan` engine
}
\keyword{internal}
