% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot.R
\name{plot_nowcast}
\alias{plot_nowcast}
\title{Create plots form the nowcast() output}
\usage{
plot_nowcast(nowcast_output)
}
\arguments{
\item{nowcast_output}{the output of the nowcast() function

# @return A a ggplot2 object that plots the evolution of the nowcast over the observed values,
it creates a subplot for each strata}
}
\description{
Plots the predictions of the function nowcast() over the input observed values
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10 (or set to iter = 2000)
now <- as.Date("1990-10-01")

# perform nowcasting
predictions <- nowcast(denguedat, "onset_week", "report_week",
  now = now, method = "optimization",
  seed = 2495624, iter = 10
)

# create summary dataframe
plot_nowcast(predictions)

}
