% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_for_nowcast.R
\name{preprocess_for_nowcast}
\alias{preprocess_for_nowcast}
\title{Preprocess data.frame for generating a nowcast}
\usage{
preprocess_for_nowcast(
  .disease_data,
  onset_date,
  report_date,
  ...,
  now,
  units,
  max_delay = Inf,
  data_type = c("auto", "linelist", "count")
)
}
\arguments{
\item{.disease_data}{A time series of reporting data in aggregated line list format
such that each row has a column for onset date, report date, and}

\item{onset_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case onset. e.g. "onset_week"}

\item{report_date}{In quotations, the name of the column of datatype
\code{Date} designating the date of case report. e.g. "report_week"}

\item{...}{Additional covariates to involve in the count.}

\item{now}{An object of datatype \code{Date} indicating the date at which
to perform the nowcast.}

\item{units}{Time scale of reporting. Options: "1 day", "1 week".}

\item{max_delay}{Maximum possible delay observed or considered for estimation
of the delay distribution (numeric). Default: `Inf`}

\item{data_type}{Either `linedata` if each row represents a test or `counts` if there
is a column named `n` with counts of how many tests had that onset and report dates}
}
\value{
A `data.frame` with all possible counts for all delay-onset combinations.
The new column with the counts is named `n`. Additional columns `.tval` and `.delay`
are added where `.tval` codifies the dates as numbers (starting at 0) and delay
codifies the difference between onset and report.
}
\description{
Function that takes a data frame with `onset_date` and `report_date` and generates all
possible combinations of onset_dates and report_dates observable controlling by the covariates
specified in `...`
}
\examples{
data(denguedat)

# Get counts by onset date and report week consider all possible delays
preprocess_for_nowcast(denguedat, "onset_week", "report_week",
  units = "weeks", now = as.Date("1990-03-05")
)

# Complete one date when there was no onset week
df <- data.frame(
  onset_week  = as.Date(c("1994-09-19", "1994-10-03", "1994-10-03", "1994-10-10")),
  report_week = as.Date(c("1994-09-19", "1994-10-03", "1994-10-10", "1994-10-10"))
)
preprocess_for_nowcast(df, "onset_week", "report_week",
  units = "weeks",
  now = as.Date("1994-10-10")
)

# Complete one date when there was no report of delay 3 mostly
df <- data.frame(
  onset_week  = as.Date(c("1994-09-19", "1994-10-03", "1994-10-03", "1994-10-10")),
  report_week = as.Date(c("1994-10-10", "1994-10-03", "1994-10-10", "1994-10-10"))
)
preprocess_for_nowcast(df, "onset_week", "report_week",
  units = "weeks",
  now = as.Date("1994-10-10")
)

# Get counts by onset date and report week stratifying by gender and state
df <- data.frame(
  onset_week = sample(as.Date(c("1994-09-19", "1994-10-03", "1994-10-10")), 100, replace = TRUE),
  gender = sample(c("Male", "Female"), 100, replace = TRUE),
  state = sample(c("A", "B", "C", "D"), prob = c(0.5, 0.2, 0.2, 0.1), size = 100, replace = TRUE)
)
df$report_week <- df$onset_week +
  sample(c(lubridate::weeks(1), lubridate::weeks(2)), 100, replace = TRUE)
preprocess_for_nowcast(df, "onset_week", "report_week", "gender", "state",
  units = "weeks",
  now = as.Date("1994-09-26")
)

}
