% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_to_numeric}
\alias{priors_to_numeric}
\title{Function for setting the priors to numeric}
\usage{
priors_to_numeric(.disease_data, priors)
}
\arguments{
\item{priors}{A list of priors generated by `set_priors()`}
}
\value{
The same list of priors but with the numeric codes RStan requires for the distributions
}
\description{
Function for setting the priors to numeric
}
\keyword{internal}
