% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_sample.R
\name{sample}
\alias{sample}
\title{Draw random samples from a prior or likelihood}
\usage{
sample(object, size, ...)
}
\arguments{
\item{object}{A \code{prior_class}, \code{likelihood_class}, or base vector.}

\item{size}{Number of draws.}

\item{...}{Passed to the likelihood method (\code{mu}, \code{phi}).}
}
\value{
A numeric/integer vector of length \code{size}.
}
\description{
Draw random samples from a prior or likelihood
}
\examples{
sample(normal_prior(log(7), 0.5), 10)
sample(gamma_prior(2, 0.1), 10)

}
