% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33_auto_nowcast.R
\name{selection_metric}
\alias{selection_metric}
\title{The metric \code{auto_nowcast()} used to pick the winner}
\usage{
selection_metric(nc)
}
\arguments{
\item{nc}{A \code{nowcast_class} returned by \code{\link[=auto_nowcast]{auto_nowcast()}}.}
}
\value{
A string: \code{"wis"}, \code{"ape"}, \code{"mse"}, \code{"coverage"}, \code{"coverage_50"},
or \code{"coverage_90"}.
}
\description{
The metric \code{auto_nowcast()} used to pick the winner
}
\seealso{
\code{\link[=auto_nowcast]{auto_nowcast()}}, \code{\link[=comparison_scores]{comparison_scores()}}
}
