% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Get nowcast priors not controlling by distribution}
\usage{
set_priors(
  mu_p = 1,
  mu_q = 1,
  nu_p = 1,
  mu_intercept_param_1 = 0,
  mu_intercept_param_2 = 1,
  mu_0_param_1 = 0,
  mu_0_param_2 = 1,
  nu_intercept_param_1 = 0,
  nu_intercept_param_2 = 1,
  nu_0_param_1 = 0,
  nu_0_param_2 = 1,
  has_cycle = FALSE,
  c_0_param_1 = 0,
  c_0_param_2 = 1,
  ctilde_0_param_1 = 0,
  ctilde_0_param_2 = 1,
  sd_mu_param_1 = 0,
  sd_mu_param_2 = 1,
  sd_nu_param_1 = 0,
  sd_nu_param_2 = 1,
  sd_c_param_1 = 0,
  sd_c_param_2 = 1,
  sd_ctilde_param_1 = 0,
  sd_ctilde_param_2 = 1,
  sd_m_param_1 = 0,
  sd_m_param_2 = 1,
  sd_dow_epi_param_1 = 0,
  sd_dow_epi_param_2 = 1,
  sd_wkend_epi_param_1 = 0,
  sd_wkend_epi_param_2 = 1,
  sd_dom_epi_param_1 = 0,
  sd_dom_epi_param_2 = 1,
  sd_month_epi_param_1 = 0,
  sd_month_epi_param_2 = 1,
  sd_week_epi_param_1 = 0,
  sd_week_epi_param_2 = 1,
  sd_holidays_epi_param_1 = 0,
  sd_holidays_epi_param_2 = 1,
  dof = 7,
  control_k_transform = 2,
  control_c_transform = 0.5
)
}
\arguments{
\item{mu_p}{Integer. Degree of the epidemic trend.}

\item{mu_q}{Integer. Degree of the errors in the epidemic trend.}

\item{nu_p}{Integer. Degree of the delay trend}

\item{mu_intercept_param_1}{Real. Mean of the epidemic intercept.}

\item{mu_intercept_param_2}{Positive real. Variance of the epidemic intercept.}

\item{mu_0_param_1}{Real. Mean of the initial value of the epidemic trend.}

\item{mu_0_param_2}{Positive Real. Variance of the initial value of the epidemic trend.}

\item{nu_intercept_param_1}{Real. Mean of the delay intercept.}

\item{nu_intercept_param_2}{Positive real. Variance of the delay intercept.}

\item{nu_0_param_1}{Real. Mean of the initial value of the disease trend.}

\item{nu_0_param_2}{Positive Real. Variance of the initial value of the disease trend.}

\item{has_cycle}{Boolean. Whether include a cycle component in the model.}

\item{c_0_param_1}{Initial value for cycle parameter's mean}

\item{c_0_param_2}{Initial value for cycle parameter's sd}

\item{ctilde_0_param_1}{Initial value for cycle parameter's mean (latent cycle)}

\item{ctilde_0_param_2}{Initial value for cycle parameter's sd (latent cycle)}

\item{sd_mu_param_1}{Positive real. Initial value for the mean of the standard deviation of the epidemic trend.}

\item{sd_mu_param_2}{Positive real. Initial value for the variance of the standard deviation of the epidemic trend.}

\item{sd_nu_param_1}{Positive real. Initial value for the mean of the standard deviation of the disease trend.}

\item{sd_nu_param_2}{Positive real. Initial value for the variance of the standard deviation of the disease trend.}

\item{sd_c_param_1}{Mean value for the standard deviation's of c_0}

\item{sd_c_param_2}{Standard deviation prior value for the standard deviation's of c_0}

\item{sd_ctilde_param_1}{Mean value for the standard deviation's of c_0 tilde}

\item{sd_ctilde_param_2}{Standard deviation prior value for the standard deviation's of c_0 tilde}

\item{sd_m_param_1}{Positive real. Initial value for the mean of the standard deviation of the observed cases (not considered if distribution is Poisson)}

\item{sd_m_param_2}{Positive real. Initial value for the variance of the standard deviation of the observed cases (not considered if distribution is Poisson)}

\item{sd_dow_epi_param_1}{Standard deviation's mean for day of the week effect}

\item{sd_dow_epi_param_2}{Standard deviation's sd for day of the week effect}

\item{sd_wkend_epi_param_1}{Standard deviation's mean for weekend effect}

\item{sd_wkend_epi_param_2}{Standard deviation's sd for weekend effect}

\item{sd_dom_epi_param_1}{Standard deviation's mean for day of the month effect}

\item{sd_dom_epi_param_2}{Standard deviation's sd for day of the month effect}

\item{sd_month_epi_param_1}{Standard deviation's mean for month effect}

\item{sd_month_epi_param_2}{Standard deviation's sd for month effect}

\item{sd_week_epi_param_1}{Standard deviation's mean for epiweek effect}

\item{sd_week_epi_param_2}{Standard deviation's sd for epiweek effect}

\item{sd_holidays_epi_param_1}{Standard deviation's mean for holiday effect}

\item{sd_holidays_epi_param_2}{Standard deviation's sd for holiday effect}

\item{dof}{Degrees of freedom if the distribution used in \code{\link[=nowcast]{nowcast()}} is Student (default = 7).}

\item{control_k_transform}{Parameter for the \code{dhyperbolic} or \code{softplus} link functions.}

\item{control_c_transform}{Parameter for the \code{dhyperbolic} link function.}
}
\value{
A list with all the priors for the \code{\link[=nowcast]{nowcast()}} function.
}
\description{
Function to return the default values for the \code{\link[=nowcast]{nowcast()}} function
as a list.
}
\examples{
#Get the default priors
set_priors()

#Change the priors
set_priors(mu_intercept_param_1 = 2, mu_q = 3)

}
