% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Get nowcast priors not controlling by distribution}
\usage{
set_priors(
  mu_p = 1,
  mu_q = 1,
  nu_p = 1,
  mu_intercept_param_1 = 0,
  mu_intercept_param_2 = 1,
  mu_0_param_1 = 0,
  mu_0_param_2 = 1,
  nu_intercept_param_1 = 0,
  nu_intercept_param_2 = 1,
  nu_0_param_1 = 0,
  nu_0_param_2 = 1,
  sd_mu_param_1 = 0,
  sd_mu_param_2 = 1,
  sd_nu_param_1 = 0,
  sd_nu_param_2 = 1,
  sd_m_param_1 = 0,
  sd_m_param_2 = 1,
  dof = 7,
  control_k_transform = 2,
  control_c_transform = 0.5
)
}
\arguments{
\item{mu_degree}{Integer. Degree of the epidemic trend.}

\item{nu_degree}{Integer. Degree of the delay trend.}

\item{p}{Integer. The number of lags to consider for an autocorrelated AR(p) model}

\item{q}{Integer. The number of lags to consider for a moving average MA(q) model}

\item{mu_is_constant}{Logical. Indicates whether the epidemic trend is constant.}

\item{nu_is_constant}{Logical. Indicates whether the delay trend is constant.}

\item{mu_sd_prior}{Character. Prior for the epidemic trend error.}

\item{nu_sd_prior}{Character. Prior for the delay trend error.}

\item{mu_sd_param_1}{Numeric. First parameter for the epidemic trend error.}

\item{mu_sd_param_2}{Numeric. Second parameter for the epidemic trend error.}

\item{nu_sd_param_1}{Numeric. First parameter for the delay trend error.}

\item{nu_sd_param_2}{Numeric. Second parameter for the delay trend error.}

\item{mu_0_mean_hyperprior}{Prior distribution for the mean of the epidemic trend.}

\item{mu_0_sd_hyperprior}{Prior distribution for the standard deviation of the epidemic trend.}

\item{mu_0_mean_param_1}{Character. Prior for the initial epidemic trend mean's first parameter.}

\item{mu_0_mean_param_2}{Character. Prior for the initial epidemic trend mean's second parameter.}

\item{mu_0_sd_param_1}{Character. Prior for the initial epidemic trend standard deviation's first parameter.}

\item{mu_0_sd_param_2}{Character. Prior for the initial epidemic trend standard deviation's second parameter.}

\item{nu_0_mean_hyperprior}{Prior distribution for the mean of the delay trend. Default is normal.}

\item{nu_0_sd_hyperprior}{Prior distribution for the standard deviation of the delay trend.}

\item{nu_0_mean_param_1}{Character. Prior for the initial delay trend mean's first parameter.}

\item{nu_0_mean_param_2}{Character. Prior for the initial delay trend mean's second parameter.}

\item{nu_0_sd_param_1}{Character. Prior for the initial delay trend standard deviation's first parameter.}

\item{nu_0_sd_param_2}{Character. Prior for the initial delay trend standard deviation's second parameter.}

\item{r_prior}{Character. Prior for the negative binomial precision parameter.}

\item{r_param_1}{Numeric. First parameter for the dispersion prior if negative binomial.}

\item{r_param_2}{Numeric. Second parameter for the dispersion prior if negative binomial.}

\item{phi_AR_param_1}{Numeric. Prior for the AR coefficients' first parameter.}

\item{phi_AR_param_2}{Numeric. Prior for the AR coefficients' second parameter.}

\item{phi_AR_prior}{String Prior distirbution name for the AR coefficients.}

\item{theta_MA_param_1}{Numeric. Prior for the MA coefficients' first parameter.}

\item{theta_MA_param_2}{Numeric. Prior for the MA coefficients' second parameter.}

\item{theta_MA_prior}{String Prior distirbution name for the MA coefficients.}

\item{xi_sd_param_1}{Numeric. Prior for the error (in log-scale) first parameter}

\item{xi_sd_param_2}{Numeric. Prior for the error (in log-scale) second parameter}

\item{xi_sd_prior}{String Prior distribution name for the errors}
}
\value{
A list with all the priors for the [nowcast()] function.
}
\description{
Function to return the default values for the [nowcast()] function
as a list.
}
\examples{
#Get the default priors
set_priors()

#Change the distribution of the priors sd for the epidemic trend as well as the degree of the delays
set_priors(nu_degree = 2, mu_sd_prior = "cauchy")

}
