% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_disease}
\alias{simulate_disease}
\title{Simulate a process}
\usage{
simulate_disease(
  num_steps = 10,
  num_delays = 8,
  num_strata = 2,
  initial_day = NULL,
  initial_n = 100,
  dist = c("NegativeBinomial", "Poisson"),
  units = c("daily", "weekly"),
  priors = set_priors()
)
}
\arguments{
\item{num_steps}{Integer. Number of time steps to simulate. Default is 10.}

\item{num_delays}{Integer. Number of delay strata. Default is 8.}

\item{num_strata}{Integer. Number of strata for the population. Default is 2.}

\item{initial_day}{Date. If the simulation is to start on a specific day.}

\item{initial_n}{Real. The initial expected average value of cases.}

\item{dist}{Distribution. Either "NegativeBinomial"  or "Poisson"}

\item{units}{Either `"daily"` (default) or `"weekly"`.}

\item{priors}{A list of all of the nowcast priors. You can use [set_priors()] to change
the priors of the function (see details)}
}
\value{
A tibble with simulated state-space process results, including the onset and
report dates, strata, delays, and the number of observed cases per time step.
}
\description{
This function simulates a state-space process for testing purposes in a model with
multiple delays and strata. It estimates the process over a spesupcified number of steps,
 with options for trend degrees, priors, and error distributions.
}
\examples{
simulate_process_for_testing(num_steps = 20, num_delays = 5, num_strata = 3)
}
