% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_process_for_testing}
\alias{simulate_process_for_testing}
\title{Simulate a process}
\usage{
simulate_process_for_testing(
  num_steps = 10,
  num_delays = 8,
  num_strata = 2,
  mu_degree = 2,
  nu_degree = 1,
  mu_is_constant = FALSE,
  nu_is_constant = TRUE,
  mu_error_prior = "normal",
  nu_error_prior = "normal",
  mu_param_1 = 0,
  mu_param_2 = 0.1,
  nu_param_1 = 0,
  nu_param_2 = 0.1,
  mu_0_prior = "normal",
  nu_0_prior = "normal",
  mu_0_param_1 = log(100),
  mu_0_param_2 = 0.01,
  nu_0_param_1 = log(6),
  nu_0_param_2 = 0.01,
  r_prior = "normal",
  is_negative_binomial = TRUE,
  r_param_1 = 0,
  r_param_2 = 1
)
}
\arguments{
\item{num_steps}{Integer. Number of time steps to simulate. Default is 10.}

\item{num_delays}{Integer. Number of delay strata. Default is 8.}

\item{num_strata}{Integer. Number of strata for the population. Default is 2.}

\item{mu_degree}{Integer. Degree of the epidemic trend. Default is 2.}

\item{nu_degree}{Integer. Degree of the delay trend. Default is 1.}

\item{mu_is_constant}{Logical. Indicates whether the epidemic trend is constant. Default is FALSE.}

\item{nu_is_constant}{Logical. Indicates whether the delay trend is constant. Default is TRUE.}

\item{mu_error_prior}{Character. Prior for the epidemic trend error. Default is "normal".}

\item{nu_error_prior}{Character. Prior for the delay trend error. Default is "normal".}

\item{mu_param_1}{Numeric. First parameter for the epidemic trend error. Default is 0.0.}

\item{mu_param_2}{Numeric. Second parameter for the epidemic trend error. Default is 1.0.}

\item{nu_param_1}{Numeric. First parameter for the delay trend error. Default is 0.0.}

\item{nu_param_2}{Numeric. Second parameter for the delay trend error. Default is 1.0.}

\item{mu_0_prior}{Character. Prior for the initial epidemic trend distribution. Default is "normal".}

\item{nu_0_prior}{Character. Prior for the initial delay trend distribution. Default is "normal".}

\item{mu_0_param_1}{Numeric. First parameter for the initial epidemic trend distribution. Default is 0.0.}

\item{mu_0_param_2}{Numeric. Second parameter for the initial epidemic trend distribution. Default is 1.0.}

\item{nu_0_param_1}{Numeric. First parameter for the initial delay trend distribution. Default is 0.0.}

\item{nu_0_param_2}{Numeric. Second parameter for the initial delay trend distribution. Default is 1.0.}

\item{r_prior}{Character. Prior for the negative binomial precision parameter. Default is "normal".}

\item{is_negative_binomial}{Logical. Indicates whether the data is modeled using a negative binomial distribution. Default is TRUE.}

\item{r_param_1}{Numeric. First parameter for the dispersion prior if negative binomial. Default is 0.0.}

\item{r_param_2}{Numeric. Second parameter for the dispersion prior if negative binomial. Default is 1.0.}
}
\value{
A tibble with simulated state-space process results, including the onset and
report dates, strata, delays, and the number of observed cases per time step.
}
\description{
This function simulates a state-space process for testing purposes in a model with
multiple delays and strata. It estimates the process over a specified number of steps,
 with options for trend degrees, priors, and error distributions.
}
\examples{
simulate_process_for_testing(num_steps = 20, num_delays = 5, num_strata = 3)
}
