% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.backtest_metrics}
\alias{summary.backtest_metrics}
\title{Makes a summary of your \code{\link[=backtest_metrics]{backtest_metrics()}} object}
\usage{
\method{summary}{backtest_metrics}(x, ..., metrics = "wis")
}
\arguments{
\item{x}{A backtest_metrics object generated by \code{\link[=backtest_metrics]{backtest_metrics()}}}

\item{...}{further arguments passed to the generic \code{\link[=summary]{summary()}} function.}

\item{metrics}{the metrics to display. Default = "wis",
accepts multiple metrics}
}
\value{
A summary tibble of the \code{\link[=backtest_metrics]{backtest_metrics()}} results,
it specifies: average, standard deviation, standard error, number of "now" dates
it also returns the metrics across all the horizons ("all" in column "horizon")
}
\description{
Makes a tidy summary dataframe of the results of the function \code{\link[=backtest_metrics]{backtest_metrics()}}
}
