% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.nowcaster}
\alias{summary.nowcaster}
\title{Makes a summary of your nowcast}
\usage{
\method{summary}{nowcaster}(object, ...)
}
\arguments{
\item{object}{A nowcaster object generated by \code{\link[=nowcast]{nowcast()}}}

\item{...}{Can be used to specify which quantiles are desired (e.g. \code{quantiles = c(0.2, 0.8)}).}
}
\value{
A summary tibble of the nowcast results,
it specifies: true date, strata, mean, standard deviation, median, and quantiles
}
\description{
Makes a tidy summary dataframe of the results of the function \code{\link[=nowcast]{nowcast()}}
}
\examples{
# Load the data
data(denguedat)

# Run a nowcast with very few iterations
# change to method = "sampling" when working and remove the iter = 10
now <- as.Date("1990-10-01")

# perform nowcasting
ncast <- nowcast(denguedat, "onset_week", "report_week",
  now = now, method = "optimization",
  seed = 2495624, iter = 10
)

# create summary dataframe
summary(ncast)

# Specific quantiles can be requested
summary(ncast, quantiles = c(0.2, 0.4))

}
